software/python/requirements.txt
crcmod==1.7
pyserial==2.7
wheel==0.24.0
